Move wxQtTreeItemEditorFactory to its own header
