+++ Makefile.in
@@ -3576,7 +3576,13 @@ COND_TOOLKIT_QT_GUI_HDR =  \
 	wx/qt/dataview.h \
 	wx/qt/dvrenderers.h \
 	$(QT_PLATFORM_HDR) \
-	wx/qt/treectrl.h
+	wx/qt/treectrl.h \
+	wx/qt/private/winevent.h \
+	wx/qt/private/converter.h \
+	wx/qt/private/treeitemfactory.h \
+	wx/qt/private/pointer.h \
+	wx/qt/private/timer.h \
+	wx/qt/private/utils.h
 @COND_TOOLKIT_QT@GUI_HDR = $(COND_TOOLKIT_QT_GUI_HDR)
 @COND_TOOLKIT_COCOA@MEDIA_PLATFORM_HDR = 
 @COND_TOOLKIT_GTK@MEDIA_PLATFORM_HDR =
